sam2/requirements.txt
eva-decord==0.6.1
hydra-core==1.3.2
jax==0.5.0
matplotlib==3.10.0
mediapipe==0.10.20
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
pillow==11.1.0
-e git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4#egg=SAM_2
scipy==1.15.1
sentencepiece==0.2.0
sounddevice==0.5.1
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1

